// src/libs/shared/items/util/models/item-types.ts
export type ItemTypes =
    'adventuringgear'
    | 'alchemicalbombs'
    | 'alchemicalelixirs'
    | 'alchemicalpoisons'
    | 'alchemicaltools'
    | 'ammunition'
    | 'armorrunes'
    | 'armors'
    | 'helditems'
    | 'materialitems'
    | 'oils'
    | 'otherconsumables'
    | 'otherconsumablesbombs'
    | 'potions'
    | 'scrolls'
    | 'shields'
    | 'snares'
    | 'talismans'
    | 'wands'
    | 'weaponrunes'
    | 'weapons'
    | 'wornitems';